# backend/openapi/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "collectZ API",
    "version": "2.5.0",
    "description": "Baseline OpenAPI contract for key auth, admin, and media endpoints."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin API server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_token",
        "description": "Browser session cookie."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "PAT or service-account key"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Required for mutating cookie-session requests."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        },
        "required": ["error"],
        "additionalProperties": true
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "email": { "type": "string", "format": "email" },
          "name": { "type": "string" },
          "role": { "type": "string" },
          "active_space_id": { "type": ["integer", "null"] },
          "active_library_id": { "type": ["integer", "null"] },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "email", "name", "role"]
      },
...
      "AuthScopeResponse": {
        "type": "object",
        "properties": {
          "active_space_id": { "type": ["integer", "null"] },
          "active_library_id": { "type": ["integer", "null"] },
          "spaces": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SpaceSummary" }
          },
          "libraries": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/LibrarySummary" }
          }
        },
        "required": ["spaces", "libraries"]
      },
      "AuthScopeSelectRequest": {
        "type": "object",
        "properties": {
          "space_id": { "type": "integer", "minimum": 1 },
          "library_id": { "type": "integer", "minimum": 1 }
        },
        "minProperties": 1
      },
...
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        },
        "required": ["email", "password"]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "user": { "$ref": "#/components/schemas/User" }
        },
        "required": ["user"]
      },
...
      "PersonalAccessTokenRecord": {
        "allOf": [
          { "$ref": "#/components/schemas/NamedTokenRecord" },
          {
            "type": "object",
            "properties": {
              "user_id": { "type": "integer" },
              "token_last_four": { "type": "string" },
              "scopes": {
                "type": "array",
                "items": { "type": "string" }
              }
            },
            "required": ["user_id", "token_last_four", "scopes"]
          }
        ]
      },
...
      "PersonalAccessTokenCreateRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "scopes": {
            "type": "array",
            "items": { "type": "string" }
          },
          "expires_at": { "type": ["string", "null"], "format": "date-time" }
        },
        "required": ["name"]
      },
      "ServiceAccountKeyCreateRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "scopes": {
            "type": "array",
            "items": { "type": "string" }
          },
          "allowed_prefixes": {
            "type": "array",
            "items": { "type": "string" }
          },
          "expires_at": { "type": ["string", "null"], "format": "date-time" }
        },
        "required": ["name"]
      },
      "PersonalAccessTokenListResponse": {
        "type": "object",
        "properties": {
          "scopes": {
            "type": "array",
            "items": { "type": "string" }
          },
          "tokens": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PersonalAccessTokenRecord" }
          }
        },
        "required": ["scopes", "tokens"]
      },
      "PersonalAccessTokenCreateResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "record": { "$ref": "#/components/schemas/PersonalAccessTokenRecord" }
        },
        "required": ["token", "record"]
      },
      "ServiceAccountKeyListResponse": {
        "type": "object",
        "properties": {
          "scopes": {
            "type": "array",
            "items": { "type": "string" }
          },
          "allowed_prefixes": {
            "type": "array",
            "items": { "type": "string" }
          },
          "keys": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ServiceAccountKeyRecord" }
          }
        },
        "required": ["scopes", "allowed_prefixes", "keys"]
      },
      "ServiceAccountKeyCreateResponse": {
        "type": "object",
        "properties": {
          "key": { "type": "string" },
          "record": { "$ref": "#/components/schemas/ServiceAccountKeyRecord" }
        },
        "required": ["key", "record"]
      },
...
    }
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Create a browser session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated user payload",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Read the current authenticated user",
        "security": [
          { "cookieSession": [] },
          { "bearerAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" }
                  },
                  "required": ["user"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/auth/scope": {
      "get": {
        "tags": ["Auth"],
        "summary": "Read the current active scope bootstrap contract",
        "security": [
          { "cookieSession": [] },
          { "bearerAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "Active auth scope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthScopeResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Auth"],
        "summary": "Select the current active space and library for the session user",
        "security": [
          { "cookieSession": [], "csrfHeader": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthScopeSelectRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated active auth scope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthScopeResponse" }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "403": {
            "description": "Scope selection denied",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/personal-access-tokens": {
      "get": {
        "tags": ["Auth"],
        "summary": "List personal access tokens for the current session user",
        "security": [
          { "cookieSession": [] }
        ],
        "responses": {
          "200": {
            "description": "Token list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PersonalAccessTokenListResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Auth"],
        "summary": "Create a personal access token",
        "security": [
          { "cookieSession": [], "csrfHeader": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PersonalAccessTokenCreateRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Token created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PersonalAccessTokenCreateResponse" }
              }
            }
          }
        }
      }
    },
    "/api/auth/personal-access-tokens/{id}": {
      "delete": {
        "tags": ["Auth"],
        "summary": "Revoke a personal access token",
        "security": [
          { "cookieSession": [], "csrfHeader": [] }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer", "minimum": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "Token revoked",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PersonalAccessTokenRecord" }
              }
            }
          },
          "404": {
            "description": "Token not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/auth/service-account-keys": {
      "get": {
        "tags": ["Auth"],
        "summary": "List service account keys",
        "security": [
          { "cookieSession": [] }
        ],
        "responses": {
          "200": {
            "description": "Service account key list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServiceAccountKeyListResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Auth"],
        "summary": "Create a service account key",
        "security": [
          { "cookieSession": [], "csrfHeader": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ServiceAccountKeyCreateRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service account key created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServiceAccountKeyCreateResponse" }
              }
            }
          }
        }
      }
    },
    "/api/auth/service-account-keys/{id}": {
      "delete": {
        "tags": ["Auth"],
        "summary": "Revoke a service account key",
        "security": [
          { "cookieSession": [], "csrfHeader": [] }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer", "minimum": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "Key revoked",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServiceAccountKeyRecord" }
              }
            }
          },
          "404": {
            "description": "Key not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  }
}